# api/internal/swagger/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: API for managing Allure test reports.
  title: AllureDeck API
  version: 2.0.0
paths:
...
  /projects/{project_id}/reports/{report_id}/stability:
    get:
      description: Returns flaky tests, regressions, and fixes for a given report.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Report ID or 'latest'
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get stability data for a report
      tags:
      - reports
...
  /projects/{project_id}/reports/{report_id}/timeline:
    get:
      description: Returns all test cases with start/stop timestamps for Gantt-chart
        rendering.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Report ID or 'latest'
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get test execution timeline for a report
      tags:
      - reports
...
  /projects/{project_id}/timeline:
    get:
      description: Returns test execution timeline data across one or more builds,
        with optional branch and date range filters.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Branch name filter
        in: query
        name: branch
        type: string
      - description: Start date (YYYY-MM-DD, inclusive)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD, inclusive end-of-day)
        in: query
        name: to
        type: string
      - description: Max builds (1-10, default 1)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get multi-build test execution timeline
      tags:
      - timeline
...
swagger: "2.0"